requirements.txt
slack-sdk==3.21.3
google-generativeai==0.3.2
gspread==5.10.0
oauth2client==4.1.3
python-dotenv==1.0.0
pytest==7.4.2
pytest-mock==3.11.1
openpyxl==3.1.2
google-api-python-client==2.100.0
google-auth==2.23.0
pytz==2023.3
